+++ xchange-binance/src/main/java/info/bitrich/xchangestream/binance/dto/TickerBinanceWebsocketTransaction.java
@@ -2,7 +2,6 @@ package info.bitrich.xchangestream.binance.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.math.BigDecimal;
 
 import org.knowm.xchange.binance.dto.marketdata.BinanceTicker24h;
 
@@ -14,21 +13,21 @@ public class TickerBinanceWebsocketTransaction extends ProductBinanceWebSocketTr
             @JsonProperty("e") String eventType,
             @JsonProperty("E") String eventTime,
             @JsonProperty("s") String symbol,
-            @JsonProperty("p") BigDecimal priceChange,
-            @JsonProperty("P") BigDecimal priceChangePercent,
-            @JsonProperty("w") BigDecimal weightedAvgPrice,
-            @JsonProperty("x") BigDecimal prevClosePrice,
-            @JsonProperty("c") BigDecimal lastPrice,
-            @JsonProperty("Q") BigDecimal lastQty,
-            @JsonProperty("b") BigDecimal bidPrice,
-            @JsonProperty("B") BigDecimal bidQty,
-            @JsonProperty("a") BigDecimal askPrice,
-            @JsonProperty("A") BigDecimal askQty,
-            @JsonProperty("o") BigDecimal openPrice,
-            @JsonProperty("h") BigDecimal highPrice,
-            @JsonProperty("l") BigDecimal lowPrice,
-            @JsonProperty("v") BigDecimal volume,
-            @JsonProperty("q") BigDecimal quoteVolume,
+            @JsonProperty("p") Double priceChange,
+            @JsonProperty("P") Double priceChangePercent,
+            @JsonProperty("w") Double weightedAvgPrice,
+            @JsonProperty("x") Double prevClosePrice,
+            @JsonProperty("c") Double lastPrice,
+            @JsonProperty("Q") Double lastQty,
+            @JsonProperty("b") Double bidPrice,
+            @JsonProperty("B") Double bidQty,
+            @JsonProperty("a") Double askPrice,
+            @JsonProperty("A") Double askQty,
+            @JsonProperty("o") Double openPrice,
+            @JsonProperty("h") Double highPrice,
+            @JsonProperty("l") Double lowPrice,
+            @JsonProperty("v") Double volume,
+            @JsonProperty("q") Double quoteVolume,
             @JsonProperty("O") Long openTime,
             @JsonProperty("C") Long closeTime,
             @JsonProperty("F") Long firstId,
