+++ xchange-bitfinex/src/main/java/info/bitrich/xchangestream/bitfinex/dto/BitfinexWebSocketAuthTrade.java
@@ -1,22 +1,21 @@
 package info.bitrich.xchangestream.bitfinex.dto;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 public class BitfinexWebSocketAuthTrade extends BitfinexWebSocketAuthPreTrade {
-    private final BigDecimal fee;
+    private final Double fee;
     private final String feeCurrency;
 
     public BitfinexWebSocketAuthTrade(long id, String pair, long mtsCreate,
-                                      long orderId, BigDecimal execAmount, BigDecimal execPrice,
-                                      String orderType, BigDecimal orderPrice,
-                                      long maker, BigDecimal fee, String feeCurrency) {
+                                      long orderId, Double execAmount, Double execPrice,
+                                      String orderType, Double orderPrice,
+                                      long maker, Double fee, String feeCurrency) {
         super(id, pair, mtsCreate, orderId, execAmount, execPrice, orderType, orderPrice, maker);
         this.fee = fee;
         this.feeCurrency = feeCurrency;
     }
 
-    public BigDecimal getFee() {
+    public Double getFee() {
         return fee;
     }
 
@@ -27,18 +26,18 @@ public class BitfinexWebSocketAuthTrade extends BitfinexWebSocketAuthPreTrade {
     @Override
     public String toString() {
         return "BitfinexWebSocketAuthenticatedTrade{" +
-            "id=" + getId() +
-            ", pair='" + getPair() + '\'' +
-            ", mtsCreate=" + getMtsCreate() +
-            ", orderId=" + getOrderId() +
-            ", execAmount=" + getExecAmount() +
-            ", execPrice=" + getExecPrice() +
-            ", orderType='" + getOrderType() + '\'' +
-            ", orderPrice=" + getOrderPrice() +
-            ", maker=" + getMtsCreate() +
-            ", fee=" + fee +
-            ", feeCurrency='" + feeCurrency + '\'' +
-            '}';
+                "id=" + getId() +
+                ", pair='" + getPair() + '\'' +
+                ", mtsCreate=" + getMtsCreate() +
+                ", orderId=" + getOrderId() +
+                ", execAmount=" + getExecAmount() +
+                ", execPrice=" + getExecPrice() +
+                ", orderType='" + getOrderType() + '\'' +
+                ", orderPrice=" + getOrderPrice() +
+                ", maker=" + getMtsCreate() +
+                ", fee=" + fee +
+                ", feeCurrency='" + feeCurrency + '\'' +
+                '}';
     }
 
     @Override
@@ -48,7 +47,7 @@ public class BitfinexWebSocketAuthTrade extends BitfinexWebSocketAuthPreTrade {
         if (!super.equals(o)) return false;
         BitfinexWebSocketAuthTrade that = (BitfinexWebSocketAuthTrade) o;
         return Objects.equals(fee, that.fee) &&
-            Objects.equals(feeCurrency, that.feeCurrency);
+                Objects.equals(feeCurrency, that.feeCurrency);
     }
 
     @Override
